chore: move types under utils
