+++ contracts/src/utils/types.ts
@@ -7,7 +7,7 @@ import {
   Signature,
   Struct,
 } from 'o1js';
-import { VALIDATOR_NUMBER } from './utils/constants';
+import { VALIDATOR_NUMBER } from './constants';
 
 export { SignaturePublicKey, SignaturePublicKeyList, List, emptyHash };
